First test pass on doing text highlighting on the GPU.
Ranges are passed into the shader using a Uniform Buffer Object. The
characters are then colored based on the color information provided in
the given range. This should be very very fast as the GPU is running
this check massively in parallel by vertex, and the lookup only needs to
happen once per range per character as it happens in the Geometry Shader
before we have generated our quads/cubes.